refactor: delete patient folder and update imports
